Send heartbeat stream to frontend
The following change adds in a new event to the eventstream that signals the request is still alive. Although the request headers contain the `keep-alive` header, this is not enough to tell a reverse proxy (or in some case browsers) that the connection is still open.

For example, if you have an nginx proxy in front of backstage the default `proxy_read_timeout` is set to 60s. If no event is sent in the eventstream during this time, NGINX cuts the request with a `(failed)net::ERR_CONNECTION_CLOSED`. We don't want this to happen.

As a way to mitigate this, I have added support to send a heartbeat event every 2 seconds. This should avoid this abrupt connection closure.

Changes to both the frontend and the backend are included in this.

Signed-off-by: Nicolas Arnold <nic@roadie.io>